+++ sdks/flutter/packages/tugboat/ios/Classes/NativeCaptureMapping.swift
@@ -1,4 +1,5 @@
 import Flutter
+import TugboatCaptureRuntime
 
 enum NativeCaptureMapping {
   static func capabilities(_ value: CaptureCapabilities) -> NativeCaptureCapabilities {